+++ Dyalect/Parser/Scanner.cs
@@ -9,8 +9,8 @@ namespace Dyalect.Parser
 {
     partial class Scanner 
     {
-	const int maxT = 68;
-	const int noSym = 68;
+	const int maxT = 67;
+	const int noSym = 67;
 
     
         static Scanner() 
@@ -21,11 +21,11 @@ namespace Dyalect.Parser
 		for (int i = 49; i <= 57; ++i) start[i] = 35;
 		start[95] = 36; 
 		start[48] = 37; 
-		start[46] = 61; 
+		start[46] = 60; 
 		start[34] = 16; 
 		start[39] = 22; 
 		start[36] = 23; 
-		start[61] = 62; 
+		start[61] = 61; 
 		start[44] = 26; 
 		start[59] = 27; 
 		start[58] = 28; 
@@ -35,19 +35,18 @@ namespace Dyalect.Parser
 		start[125] = 32; 
 		start[91] = 33; 
 		start[93] = 34; 
-		start[43] = 63; 
-		start[45] = 64; 
-		start[42] = 65; 
-		start[47] = 66; 
-		start[37] = 67; 
-		start[124] = 68; 
-		start[38] = 69; 
-		start[33] = 70; 
-		start[35] = 41; 
-		start[62] = 71; 
-		start[60] = 72; 
-		start[94] = 73; 
-		start[126] = 46; 
+		start[43] = 62; 
+		start[45] = 63; 
+		start[42] = 64; 
+		start[47] = 65; 
+		start[37] = 66; 
+		start[124] = 67; 
+		start[38] = 68; 
+		start[33] = 69; 
+		start[62] = 70; 
+		start[60] = 71; 
+		start[94] = 72; 
+		start[126] = 45; 
 		start[Buffer.EOF] = -1;
 
         }
@@ -151,12 +150,12 @@ namespace Dyalect.Parser
 			case "for": t.kind = 15; break;
 			case "while": t.kind = 16; break;
 			case "type": t.kind = 17; break;
-			case "import": t.kind = 49; break;
-			case "else": t.kind = 51; break;
-			case "in": t.kind = 52; break;
-			case "nil": t.kind = 65; break;
-			case "true": t.kind = 66; break;
-			case "false": t.kind = 67; break;
+			case "import": t.kind = 48; break;
+			case "else": t.kind = 50; break;
+			case "in": t.kind = 51; break;
+			case "nil": t.kind = 64; break;
+			case "true": t.kind = 65; break;
+			case "false": t.kind = 66; break;
 			default: break;
 		}
         }
@@ -343,18 +342,18 @@ namespace Dyalect.Parser
 			case 42:
 				{t.kind = 39; break;}
 			case 43:
-				{t.kind = 40; break;}
+				{t.kind = 42; break;}
 			case 44:
 				{t.kind = 43; break;}
 			case 45:
-				{t.kind = 44; break;}
+				{t.kind = 47; break;}
 			case 46:
-				{t.kind = 48; break;}
-			case 47:
-				if (ch == '.') {AddCh(); goto case 48;}
+				if (ch == '.') {AddCh(); goto case 47;}
 				else {goto case 0;}
+			case 47:
+				{t.kind = 49; break;}
 			case 48:
-				{t.kind = 50; break;}
+				{t.kind = 52; break;}
 			case 49:
 				{t.kind = 53; break;}
 			case 50:
@@ -378,73 +377,71 @@ namespace Dyalect.Parser
 			case 59:
 				{t.kind = 63; break;}
 			case 60:
-				{t.kind = 64; break;}
-			case 61:
 				recEnd = pos; recKind = 19;
 				if (ch >= '0' && ch <= '9') {AddCh(); goto case 4;}
-				else if (ch == '.') {AddCh(); goto case 47;}
+				else if (ch == '.') {AddCh(); goto case 46;}
 				else {t.kind = 19; break;}
-			case 62:
+			case 61:
 				recEnd = pos; recKind = 23;
 				if (ch == '>') {AddCh(); goto case 25;}
-				else if (ch == '=') {AddCh(); goto case 42;}
+				else if (ch == '=') {AddCh(); goto case 41;}
 				else {t.kind = 23; break;}
-			case 63:
+			case 62:
 				recEnd = pos; recKind = 30;
-				if (ch == '=') {AddCh(); goto case 49;}
+				if (ch == '=') {AddCh(); goto case 48;}
 				else {t.kind = 30; break;}
-			case 64:
+			case 63:
 				recEnd = pos; recKind = 31;
-				if (ch == '=') {AddCh(); goto case 50;}
+				if (ch == '=') {AddCh(); goto case 49;}
 				else {t.kind = 31; break;}
-			case 65:
+			case 64:
 				recEnd = pos; recKind = 32;
-				if (ch == '=') {AddCh(); goto case 51;}
+				if (ch == '=') {AddCh(); goto case 50;}
 				else {t.kind = 32; break;}
-			case 66:
+			case 65:
 				recEnd = pos; recKind = 33;
-				if (ch == '=') {AddCh(); goto case 52;}
+				if (ch == '=') {AddCh(); goto case 51;}
 				else {t.kind = 33; break;}
-			case 67:
+			case 66:
 				recEnd = pos; recKind = 34;
-				if (ch == '=') {AddCh(); goto case 53;}
+				if (ch == '=') {AddCh(); goto case 52;}
 				else {t.kind = 34; break;}
-			case 68:
+			case 67:
 				recEnd = pos; recKind = 35;
-				if (ch == '=') {AddCh(); goto case 55;}
-				else if (ch == '|') {AddCh(); goto case 59;}
-				else {t.kind = 35; break;}
-			case 69:
-				recEnd = pos; recKind = 36;
 				if (ch == '=') {AddCh(); goto case 54;}
-				else if (ch == '&') {AddCh(); goto case 60;}
+				else if (ch == '|') {AddCh(); goto case 58;}
+				else {t.kind = 35; break;}
+			case 68:
+				recEnd = pos; recKind = 36;
+				if (ch == '=') {AddCh(); goto case 53;}
+				else if (ch == '&') {AddCh(); goto case 59;}
 				else {t.kind = 36; break;}
-			case 70:
+			case 69:
 				recEnd = pos; recKind = 37;
-				if (ch == '=') {AddCh(); goto case 43;}
+				if (ch == '=') {AddCh(); goto case 42;}
 				else {t.kind = 37; break;}
+			case 70:
+				recEnd = pos; recKind = 40;
+				if (ch == '=') {AddCh(); goto case 43;}
+				else if (ch == '>') {AddCh(); goto case 73;}
+				else {t.kind = 40; break;}
 			case 71:
 				recEnd = pos; recKind = 41;
 				if (ch == '=') {AddCh(); goto case 44;}
-				else if (ch == '>') {AddCh(); goto case 74;}
+				else if (ch == '<') {AddCh(); goto case 74;}
 				else {t.kind = 41; break;}
 			case 72:
-				recEnd = pos; recKind = 42;
-				if (ch == '=') {AddCh(); goto case 45;}
-				else if (ch == '<') {AddCh(); goto case 75;}
-				else {t.kind = 42; break;}
+				recEnd = pos; recKind = 44;
+				if (ch == '=') {AddCh(); goto case 55;}
+				else {t.kind = 44; break;}
 			case 73:
-				recEnd = pos; recKind = 45;
-				if (ch == '=') {AddCh(); goto case 56;}
-				else {t.kind = 45; break;}
-			case 74:
-				recEnd = pos; recKind = 47;
-				if (ch == '=') {AddCh(); goto case 58;}
-				else {t.kind = 47; break;}
-			case 75:
 				recEnd = pos; recKind = 46;
 				if (ch == '=') {AddCh(); goto case 57;}
 				else {t.kind = 46; break;}
+			case 74:
+				recEnd = pos; recKind = 45;
+				if (ch == '=') {AddCh(); goto case 56;}
+				else {t.kind = 45; break;}
 
             }