+++ Dyalect/Runtime/Types/DyString.cs
@@ -118,6 +118,8 @@ namespace Dyalect.Runtime.Types
 
         protected override DyFunction GetTrait(string name, ExecutionContext ctx)
         {
+            if (name == Builtins.Len)
+                return DyForeignFunction.Create(name, LenAdapter);
 
             return null;
         }